Added support for mnemonic tokens that require a registry (typed directly or by selector code)

+++ inc/Assembler/Statement.h
@@ -8,7 +8,7 @@
 struct Statement
 {
 	uint8_t opcode = 0x00;
-	size_t operatorCount = 0u;
+	uint8_t operatorCount = 0u;
 	std::array<uint8_t, 2> operands{};
 	const ISA::ISAEntry* ISAEntry = nullptr; // temporary, for print debugging
