+++ src/Utils/StringUtils.cpp
@@ -43,7 +43,7 @@ bool Utils::IsNumber(std::string_view word)
 		});
 }
 
-uint16_t Utils::ParseNumber(const std::string& word)
+uint32_t Utils::ParseNumber(const std::string& word)
 {
 	uint8_t base = 10u;
 	if (HasPrefix(word))
@@ -51,7 +51,7 @@ uint16_t Utils::ParseNumber(const std::string& word)
 		const std::string prefix = word.substr(0, 2);
 		CheckBase(prefix, base);
 	}
-	return static_cast<uint16_t>(std::stoi(word, nullptr, base));
+	return static_cast<uint32_t>(std::stoi(word, nullptr, base));
 }
 
 std::string_view Utils::OperatorKindToString(ISA::OperatorKind operatorKind)
